Further refactoring, added Notification Repository
