+++ ArtAttack/Repository/NotificationRepository.cs
@@ -1,26 +1,28 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 using ArtAttack.Domain;
-using ArtAttack.Model;
 using ArtAttack.Shared;
 
-namespace ArtAttack.Model
+namespace ArtAttack.Repository
 {
-    public class NotificationDataAdapter : IDisposable, INotificationDataAdapter
+    public class NotificationRepository : INotificationRepository
     {
         private readonly IDatabaseProvider databaseProvider;
         private readonly string connectionString;
         private IDbConnection connection;
 
         [ExcludeFromCodeCoverage]
-        public NotificationDataAdapter(string connectionString)
+        public NotificationRepository(string connectionString)
             : this(connectionString, new SqlDatabaseProvider())
         {
         }
 
-        public NotificationDataAdapter(string connectionString, IDatabaseProvider databaseProvider)
+        public NotificationRepository(string connectionString, IDatabaseProvider databaseProvider)
         {
             if (connectionString == null)
             {
@@ -62,7 +64,7 @@ namespace ArtAttack.Model
                 {
                     while (reader.Read())
                     {
-                        notifications.Add(NotificationFactory.CreateFromDataReader(reader));
+                        notifications.Add(this.CreateFromDataReader(reader));
                     }
                 }
             }
@@ -217,6 +219,68 @@ namespace ArtAttack.Model
         {
             connection?.Dispose();
             connection = null;
+        }
+
+        /// <summary>
+        /// Creates a Notification object from a data reader
+        /// </summary>
+        /// <param name="reader">The reader to create a notification from</param>
+        /// <returns></returns>
+        /// <exception cref="ArgumentException"></exception>
+        public Notification CreateFromDataReader(IDataReader reader)
+        {
+            int notificationId = reader.GetInt32(reader.GetOrdinal("notificationID"));
+            int recipientId = reader.GetInt32(reader.GetOrdinal("recipientID"));
+            DateTime timestamp = reader.GetDateTime(reader.GetOrdinal("timestamp"));
+            bool isRead = reader.GetBoolean(reader.GetOrdinal("isRead"));
+            string category = reader.GetString(reader.GetOrdinal("category"));
+
+            switch (category)
+            {
+                case "CONTRACT_RENEWAL_ACCEPTED":
+                    int contractId = reader.GetInt32(reader.GetOrdinal("contractID"));
+                    bool isAccepted = reader.GetBoolean(reader.GetOrdinal("isAccepted"));
+                    return new ContractRenewalAnswerNotification(recipientId, timestamp, contractId, isAccepted, isRead, notificationId);
+
+                case "CONTRACT_RENEWAL_WAITLIST":
+                    int productIdWaitlist = reader.GetInt32(reader.GetOrdinal("productID"));
+                    return new ContractRenewalWaitlistNotification(recipientId, timestamp, productIdWaitlist, isRead, notificationId);
+
+                case "OUTBIDDED":
+                    int productIdOutbidded = reader.GetInt32(reader.GetOrdinal("productID"));
+                    return new OutbiddedNotification(recipientId, timestamp, productIdOutbidded, isRead, notificationId);
+
+                case "ORDER_SHIPPING_PROGRESS":
+                    int orderId = reader.GetInt32(reader.GetOrdinal("orderID"));
+                    string shippingState = reader.GetString(reader.GetOrdinal("shippingState"));
+                    DateTime deliveryDate = reader.GetDateTime(reader.GetOrdinal("deliveryDate"));
+                    return new OrderShippingProgressNotification(recipientId, timestamp, orderId, shippingState, deliveryDate, isRead, notificationId);
+
+                case "PAYMENT_CONFIRMATION":
+                    int productIdPayment = reader.GetInt32(reader.GetOrdinal("productID"));
+                    int orderIdPayment = reader.GetInt32(reader.GetOrdinal("orderID"));
+                    return new PaymentConfirmationNotification(recipientId, timestamp, productIdPayment, orderIdPayment, isRead, notificationId);
+
+                case "PRODUCT_REMOVED":
+                    int productIdRemoved = reader.GetInt32(reader.GetOrdinal("productID"));
+                    return new ProductRemovedNotification(recipientId, timestamp, productIdRemoved, isRead, notificationId);
+
+                case "PRODUCT_AVAILABLE":
+                    int productIdAvailable = reader.GetInt32(reader.GetOrdinal("productID"));
+                    return new ProductAvailableNotification(recipientId, timestamp, productIdAvailable, isRead, notificationId);
+
+                case "CONTRACT_RENEWAL_REQUEST":
+                    int contractIdReq = reader.GetInt32(reader.GetOrdinal("contractID"));
+                    return new ContractRenewalRequestNotification(recipientId, timestamp, contractIdReq, isRead, notificationId);
+
+                case "CONTRACT_EXPIRATION":
+                    int contractIdExp = reader.GetInt32(reader.GetOrdinal("contractID"));
+                    DateTime expirationDate = reader.GetDateTime(reader.GetOrdinal("expirationDate"));
+                    return new ContractExpirationNotification(recipientId, timestamp, contractIdExp, expirationDate, isRead, notificationId);
+
+                default:
+                    throw new ArgumentException($"Unknown notification category: {category}");
+            }
         }
     }
 }
