+++ ArtAttack/Shared/Configuration.cs
@@ -2,6 +2,6 @@
 {
     public class Configuration
     {
-        public static string CONNECTION_STRING = "Data Source=RAZVAN\\SQLEXPRESS01;Initial Catalog=PurchaseDatabase;Integrated Security=True;Connect Timeout=30;Encrypt=True;Trust Server Certificate=True;Application Intent=ReadWrite;Multi Subnet Failover=False";
+        public static string CONNECTION_STRING = "Data Source=DESKTOP-SAGPO35\\SQLEXPRESS;Initial Catalog=ArtAttack;Integrated Security=True;Connect Timeout=30;Encrypt=True;Trust Server Certificate=True;Application Intent=ReadWrite;Multi Subnet Failover=False";
     }
 }
